requirements.txt
openai>=1.0.0
chromadb>=0.4.0
flask>=3.0.0
flask-socketio>=5.3.0
docker>=7.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
eventlet>=0.33.0

